backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
kubernetes==35.0.0
pydantic==2.12.5
pyyaml==6.0.3
python-multipart==0.0.22
aiofiles==25.1.0